Use @scandipwa/scandipwa-extensibility Migrate configurations from next.config.js Ignore the new stuff from ESLint
