--- next-packages/nextjs-extensibility/lib/handlers/generateApplyHandler.js
@@ -1,21 +1,0 @@
-import getPluginsForMember from '../helpers/getPluginsForMember';
-import generateMiddlewaredFunction from '../middlewarers/generateMiddlewaredFunction';
-
-export const withNamespace = (namespace) => (origFunction, thisArg, originalArgs) => {
-    // Get plugins for the function
-    const memberPluginsApply = getPluginsForMember(namespace, 'function');
-
-    // If no plugins => return the original function
-    if (!memberPluginsApply.length) {
-        return origFunction.apply(thisArg, originalArgs);
-    }
-
-    // Return the result of a call of a generated function (=wrapped into plugins)
-    return generateMiddlewaredFunction(
-        origFunction,
-        memberPluginsApply,
-        thisArg
-    )(...originalArgs);
-};
-
-export default withNamespace;
