--- next-packages/nextjs-extensibility/lib/middlewarers/generateMiddlewaredFunction.js
@@ -1,30 +1,0 @@
-import getWrapperFromPlugin from '../helpers/getWrapperFromPlugin';
-
-/**
- * Middlewaring given original member
- * @param {Function} origMember
- * @param {Array} sortedPlugins
- * @param Context origContext
- */
-export const withNamespace = (origMember = () => {}, sortedPlugins, origContext) => (...args) => {
-    const newMember = sortedPlugins.reduce(
-        (acc, plugin) => () => {
-            const wrapper = getWrapperFromPlugin(plugin, origMember.name);
-
-            // Provide different arguments due to API difference
-            // Between property and function call interception
-            return typeof origMember === 'object'
-                ? wrapper(acc, origContext)
-                : wrapper(
-                    args,
-                    acc.bind(origContext),
-                    origContext
-                );
-        },
-        origMember
-    );
-
-    return newMember(args);
-};
-
-export default withNamespace;
