--- next-packages/nextjs-extensibility/package.json
@@ -1,20 +1,0 @@
-{
-    "name": "@scandipwa/nextjs-extensibility",
-    "description": "Extension logic used by CSA.",
-    "version": "0.0.0",
-    "license": "OSL-3.0",
-    "publishConfig": {
-        "access": "public"
-    },
-    "scandipwa": {
-        "type": "extension"
-    },
-    "dependencies": {
-        "@babel/plugin-transform-arrow-functions": "^7.10.4",
-        "@babel/plugin-transform-async-to-generator": "^7.10.4",
-        "@scandipwa/craco": "0.1.1",
-        "@scandipwa/scandipwa-dev-utils": "0.0.20",
-        "noop-loader": "^1.0.0",
-        "null-loader": "^4.0.1"
-    }
-}
