+++ next-packages/nextjs-scripts/package.json
@@ -5,6 +5,12 @@
     "bin": {
         "nextjs-scripts": "./bin/nextjs-scripts.js"
     },
+    "scandipwa": {
+        "extensions": {
+            "@scandipwa/scandipwa-extensibility": true,
+            "@scandipwa/nextjs-scripts": true
+        }
+    },
     "engines": {
         "node": ">=12.0.0"
     },
