+++ next-packages/scandipwa-extensibility/LICENSE
