+++ next-packages/scandipwa-extensibility/README.md
