+++ next-packages/scandipwa-extensibility/build-config/config.plugin.js
@@ -1,0 +1,13 @@
+const injectBabelConfig = require('./util/babel');
+const injectWebpackConfig = require('./util/webpack');
+
+module.exports = {
+    plugin: {
+        overrideCracoConfig: ({ cracoConfig }) => {
+            injectBabelConfig(cracoConfig.babel);
+
+            return cracoConfig;
+        },
+        overrideWebpackConfig: ({ webpackConfig }) => injectWebpackConfig(webpackConfig)
+    }
+};
