+++ next-packages/scandipwa-extensibility/build-config/util/babel/index.js
@@ -1,0 +1,5 @@
+const addPlugins = require('./add-plugins');
+
+const injectBabelConfig = (babelConfig) => addPlugins(babelConfig);
+
+module.exports = injectBabelConfig;
