+++ next-packages/scandipwa-extensibility/build-config/util/common/safe-path.js
@@ -1,0 +1,5 @@
+
+const PATH_DELIMITER = '[\\\\/]'; // match 2 antislashes or one slash
+const safePath = (module) => module.split('/').join(PATH_DELIMITER);
+
+module.exports = safePath;
