+++ next-packages/scandipwa-extensibility/build-config/util/index.js
@@ -1,0 +1,6 @@
+const injectBabelConfig = require('./babel');
+const injectWebpackConfig = require('./webpack');
+
+const configInjector = { injectBabelConfig, injectWebpackConfig }
+
+module.exports = configInjector;
