+++ next-packages/scandipwa-extensibility/build-config/util/webpack/inject-loader.js
@@ -1,0 +1,11 @@
+// Inject the actual extensions' imports
+const addImportInjectorLoader = (webpackConfig, entryMatcher) => {
+    webpackConfig.module.rules.push({
+        test: entryMatcher || webpackConfig.entry,
+        loader: require.resolve('../../webpack-extension-import-loader')
+    });
+
+    return webpackConfig;
+};
+
+module.exports = addImportInjectorLoader;
