+++ next-packages/scandipwa-extensibility/lib/extensible/constants.js
@@ -1,0 +1,5 @@
+export class EmptyBase {}
+
+// Key for processable classes to determine whether they are already extensible
+export const extensible = Symbol('Extensible');
+export const cacheIdentityKey = Symbol('CacheIdentityKey');
