+++ next-packages/scandipwa-extensibility/lib/namespace/get-namespaces-from-middlewarable.js
@@ -1,0 +1,3 @@
+export default function getNamespacesFromMiddlewarable(Middlewarable) {
+    return Middlewarable.prototype.__namespaces__;
+}
