+++ next-packages/scandipwa-extensibility/lib/plugins/get-wrapper-from-plugin.js
@@ -1,0 +1,15 @@
+export default function getWrapperFromPlugin(plugin, memberName) {
+    const { implementation } = plugin;
+
+    if (typeof plugin === 'function') {
+        return plugin;
+    }
+
+    if (typeof implementation === 'function') {
+        return implementation;
+    }
+
+    throw new Error(
+        `No implementation found in plugin definition for member ${memberName}.`
+    );
+}
