+++ next-packages/scandipwa-extensibility/lib/plugins/plugin-storage.js
@@ -1,0 +1,19 @@
+import getGlobalContext from '../context/get-global-context';
+import generateConfig from './generate-config';
+
+class PluginStorage {
+    constructor() {
+        // Plugins' storage, ex: window.plugins
+        this.plugins = [];
+    }
+
+    setPlugins(importArray) {
+        this.plugins = generateConfig(importArray);
+
+        // Expose this.plugins to the global context
+        // For debugging convenience; this is not used in the code
+        getGlobalContext().plugins = this.plugins;
+    }
+}
+
+export default new PluginStorage();
