+++ next-packages/scandipwa-extensibility/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "@scandipwa/scandipwa-extensibility",
+    "description": "Extension logic used by CSA.",
+    "version": "2.0.2",
+    "license": "OSL-3.0",
+    "types": "./typings",
+    "publishConfig": {
+        "access": "public"
+    },
+    "scandipwa": {
+        "type": "extension",
+        "build": {
+            "cracoPlugins": [
+                "build-config/config.plugin.js"
+            ]
+        }
+    },
+    "dependencies": {
+        "@babel/plugin-transform-arrow-functions": "^7.10.4",
+        "@babel/plugin-transform-async-to-generator": "^7.10.4",
+        "@scandipwa/craco": "0.1.1",
+        "@scandipwa/scandipwa-dev-utils": "0.0.20",
+        "noop-loader": "^1.0.0",
+        "null-loader": "^4.0.1"
+    }
+}
